refactor parser modules into parser package
