+++ src/functional/ffront/parsing/field_operator_parser.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-#
 # GT4Py Project - GridTools Framework
 #
 # Copyright (c) 2014-2021, ETH Zurich
@@ -16,33 +14,10 @@
 
 import ast
 import inspect
-import textwrap
-from typing import Callable, List, Optional
+from typing import Callable, List
 
-from eve import NodeTranslator
 from functional.ffront import field_operator_ir as foir
-from functional.iterator import ir as iir
-
-
-class EmptyReturnException(Exception):
-    ...
-
-
-class FieldOperatorSyntaxError(SyntaxError):
-    def __init__(self, msg="", *, lineno=0, offset=0, filename=None):
-        self.msg = "Invalid Field Operator Syntax: " + msg
-        self.lineno = lineno
-        self.offset = offset
-        self.filename = filename
-
-
-def get_ast_from_func(func: Callable) -> ast.stmt:
-    if inspect.getabsfile(func) == "<string>":
-        raise ValueError(
-            "Can not create field operator from a function that is not in a source file!"
-        )
-    source = textwrap.dedent(inspect.getsource(func))
-    return ast.parse(source).body[0]
+from functional.ffront.parsing.parsing import FieldOperatorSyntaxError, get_ast_from_func
 
 
 class FieldOperatorParser(ast.NodeVisitor):
@@ -111,58 +86,3 @@ class FieldOperatorParser(ast.NodeVisitor):
             lineno=node.lineno,
             offset=node.col_offset,
         )
-
-
-class SymExprResolver(NodeTranslator):
-    @classmethod
-    def parse(cls, nodes: List[foir.Expr], *, params: Optional[list[iir.Sym]] = None) -> foir.Expr:
-        names: dict[str, foir.Expr] = {}
-        parser = cls()
-        for node in nodes[:-1]:
-            names.update(parser.visit(node, names=names))
-        return foir.Return(value=parser.visit(nodes[-1], names=names))
-
-    def visit_SymExpr(
-        self,
-        node: foir.SymExpr,
-        *,
-        names: Optional[dict[str, foir.Expr]] = None,
-    ) -> dict[str, iir.Expr]:
-        return {node.id: self.visit(node.expr, names=names)}
-
-    def visit_Name(
-        self,
-        node: foir.Name,
-        *,
-        names: Optional[dict[str, foir.Expr]] = None,
-    ):
-        names = names or {}
-        if node.id in names:
-            return names[node.id]
-        return foir.SymRef(id=node.id)
-
-
-class FieldOperatorLowering(NodeTranslator):
-    @classmethod
-    def parse(cls, node: foir.FieldOperator) -> iir.FunctionDefinition:
-        return cls().visit(node)
-
-    def visit_FieldOperator(self, node: foir.FieldOperator) -> iir.FunctionDefinition:
-        params = self.visit(node.params)
-        return iir.FunctionDefinition(
-            id=node.id, params=params, expr=self.body_visit(node.body, params=params)
-        )
-
-    def body_visit(
-        self, exprs: List[foir.Expr], params: Optional[List[iir.Sym]] = None
-    ) -> iir.Expr:
-        return self.visit(SymExprResolver.parse(exprs))
-
-    def visit_Return(self, node: foir.Return) -> iir.Expr:
-        return self.visit(node.value)
-
-    def visit_Sym(self, node: foir.Sym) -> iir.Sym:
-        return iir.Sym(id=node.id)
-
-    def visit_SymRef(self, node: foir.SymRef) -> iir.FunCall:
-        return iir.FunCall(fun=iir.SymRef(id="deref"), args=[iir.SymRef(id=node.id)])
